% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractr.R
\name{check_retraction}
\alias{check_retraction}
\title{Check Retraction Status}
\usage{
check_retraction(doi = c())
}
\arguments{
\item{doi}{a vector of DOIs.}
}
\value{
a data.frame of 6 columns; \code{retracted}, \code{title}, \code{url},
\code{journal}, \code{update} and \code{doi}. In the event that the paper has not
been retracted, all the metadata will be \code{NA}, while \code{retracted} will
be FALSE.
}
\description{
\code{check_retraction} tests whether a paper (or set of papers)
have been retracted, accepting a set of DOIs and returning the retraction
status (as a boolean) and paper metadata in the case it has been retracted.
}
\examples{

\dontrun{
# check for retractions
check_retraction(doi = "10.7860/JCDR/2013/4833.2724")
}
}
