% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractr.R
\docType{package}
\name{retractr}
\alias{retractr}
\alias{retractr-package}
\alias{retractr-package}
\title{Open Retractions client}
\description{
This package provides access to the Open Retractions database,
which contains information on retracted papers, through its programmatic API.
}
\seealso{
\code{\link{check_retraction}}, the package's sole function.
}
